Rename EventSource and StateHolder
