+++ feature/login/src/main/java/com/epmedu/animeal/login/phone/domain/EnterPhoneViewModel.kt
@@ -3,8 +3,8 @@ package com.epmedu.animeal.login.phone.domain
 import android.telephony.PhoneNumberUtils
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.epmedu.animeal.common.presentation.state.StateHolder
-import com.epmedu.animeal.common.presentation.state.StateHolderImpl
+import com.epmedu.animeal.common.presentation.viewmodel.delegate.DefaultStateDelegate
+import com.epmedu.animeal.common.presentation.viewmodel.delegate.StateDelegate
 import com.epmedu.animeal.login.phone.data.EnterPhoneRepository
 import com.epmedu.animeal.login.phone.domain.model.EnterPhoneState
 import dagger.hilt.android.lifecycle.HiltViewModel
@@ -15,7 +15,7 @@ import javax.inject.Inject
 internal class EnterPhoneViewModel @Inject constructor(
     private val repository: EnterPhoneRepository
 ) : ViewModel(),
-    StateHolder<EnterPhoneState> by StateHolderImpl(initialState = EnterPhoneState()) {
+    StateDelegate<EnterPhoneState> by DefaultStateDelegate(initialState = EnterPhoneState()) {
 
     fun updatePhoneNumber(newNumber: String) {
         updateState {
