+++ feature/login/src/main/java/com/epmedu/animeal/login/profile/presentation/FinishProfileViewModel.kt
@@ -4,10 +4,10 @@ import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.epmedu.animeal.common.data.model.Profile
 import com.epmedu.animeal.common.data.repository.ProfileRepository
-import com.epmedu.animeal.common.presentation.event.EventSource
-import com.epmedu.animeal.common.presentation.event.EventSourceImpl
-import com.epmedu.animeal.common.presentation.state.StateHolder
-import com.epmedu.animeal.common.presentation.state.StateHolderImpl
+import com.epmedu.animeal.common.presentation.viewmodel.delegate.DefaultEventDelegate
+import com.epmedu.animeal.common.presentation.viewmodel.delegate.DefaultStateDelegate
+import com.epmedu.animeal.common.presentation.viewmodel.delegate.EventDelegate
+import com.epmedu.animeal.common.presentation.viewmodel.delegate.StateDelegate
 import com.epmedu.animeal.extensions.DAY_MONTH_COMMA_YEAR_FORMATTER
 import com.epmedu.animeal.extensions.formatDateToString
 import com.epmedu.animeal.foundation.common.validation.ProfileValidator
@@ -20,8 +20,8 @@ import javax.inject.Inject
 internal class FinishProfileViewModel @Inject constructor(
     private val profileRepository: ProfileRepository
 ) : ViewModel(),
-    StateHolder<FinishProfileState> by StateHolderImpl(initialState = FinishProfileState()),
-    EventSource<Event> by EventSourceImpl() {
+    StateDelegate<FinishProfileState> by DefaultStateDelegate(initialState = FinishProfileState()),
+    EventDelegate<Event> by DefaultEventDelegate() {
 
     private val validator: ProfileValidator = ProfileValidator()
