+++ feature/more/src/main/java/com/epmedu/animeal/more/help/HelpViewModel.kt
@@ -2,8 +2,8 @@ package com.epmedu.animeal.more.help
 
 import androidx.lifecycle.ViewModel
 import com.epmedu.animeal.common.component.BuildConfigProvider
-import com.epmedu.animeal.common.presentation.state.StateHolder
-import com.epmedu.animeal.common.presentation.state.StateHolderImpl
+import com.epmedu.animeal.common.presentation.viewmodel.delegate.DefaultStateDelegate
+import com.epmedu.animeal.common.presentation.viewmodel.delegate.StateDelegate
 import dagger.hilt.android.lifecycle.HiltViewModel
 import javax.inject.Inject
 
@@ -11,7 +11,7 @@ import javax.inject.Inject
 internal class HelpViewModel @Inject constructor(
     private val buildConfigProvider: BuildConfigProvider
 ) : ViewModel(),
-    StateHolder<HelpState> by StateHolderImpl(initialState = HelpState()) {
+    StateDelegate<HelpState> by DefaultStateDelegate(initialState = HelpState()) {
 
     init {
         updateState {
