+++ feature/more/src/main/java/com/epmedu/animeal/more/profile/ProfileViewModel.kt
@@ -4,8 +4,8 @@ import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.epmedu.animeal.common.data.model.Profile
 import com.epmedu.animeal.common.data.repository.ProfileRepository
-import com.epmedu.animeal.common.presentation.state.StateHolder
-import com.epmedu.animeal.common.presentation.state.StateHolderImpl
+import com.epmedu.animeal.common.presentation.viewmodel.delegate.DefaultStateDelegate
+import com.epmedu.animeal.common.presentation.viewmodel.delegate.StateDelegate
 import com.epmedu.animeal.extensions.DAY_MONTH_COMMA_YEAR_FORMATTER
 import com.epmedu.animeal.extensions.formatDateToString
 import com.epmedu.animeal.foundation.common.validation.ProfileValidator
@@ -18,7 +18,7 @@ import javax.inject.Inject
 internal class ProfileViewModel @Inject constructor(
     private val profileRepository: ProfileRepository
 ) : ViewModel(),
-    StateHolder<ProfileState> by StateHolderImpl(initialState = ProfileState()) {
+    StateDelegate<ProfileState> by DefaultStateDelegate(initialState = ProfileState()) {
 
     private val validator: ProfileValidator = ProfileValidator()
