+++ feature/splash/src/main/java/com/epmedu/animeal/splash/presentation/SplashViewModel.kt
@@ -3,8 +3,8 @@ package com.epmedu.animeal.splash.presentation
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.epmedu.animeal.common.data.repository.ProfileRepository
-import com.epmedu.animeal.common.presentation.event.EventSource
-import com.epmedu.animeal.common.presentation.event.EventSourceImpl
+import com.epmedu.animeal.common.presentation.viewmodel.delegate.DefaultEventDelegate
+import com.epmedu.animeal.common.presentation.viewmodel.delegate.EventDelegate
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.launch
 import javax.inject.Inject
@@ -13,7 +13,7 @@ import javax.inject.Inject
 class SplashViewModel @Inject constructor(
     private val profileRepository: ProfileRepository
 ) : ViewModel(),
-    EventSource<SplashViewModel.Event> by EventSourceImpl() {
+    EventDelegate<SplashViewModel.Event> by DefaultEventDelegate() {
 
     fun verifyProfileSaved() {
         viewModelScope.launch {
